Migrate Express to nextjs app
