--- agent-tee-phala/.env.example
@@ -1,3 +1,0 @@
-TEE_SERVER_URL=http://localhost:8090
-TEE_SERVER_PORT=8090
-CROSSMINT_SERVER_API_KEY=
