--- agent-tee-phala/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "agent-tee-phala",
-  "version": "1.0.0",
-  "scripts": {
-    "build": "tsc",
-    "start": "node dist/index.js",
-    "dev": "ts-node-dev --respawn src/index.ts",
-    "format": "biome format --write"
-  },
-  "dependencies": {
-    "@phala/dstack-sdk": "latest",
-    "express": "^4.18.2",
-    "cors": "^2.8.5",
-    "dotenv": "^16.0.3",
-    "viem": "2.22.11"
-  },
-  "devDependencies": {
-    "@biomejs/biome": "^1.9.4",
-    "@types/express": "^4.17.21",
-    "@types/cors": "^2.8.17",
-    "@types/node": "^20.0.0",
-    "typescript": "^5.0.0",
-    "ts-node-dev": "^2.0.0"
-  }
-}
