--- agent-tee-phala/src/types/index.ts
@@ -1,10 +1,0 @@
-export interface DerivedKeys {
-    key: string;
-    timestamp: string;
-}
-
-export interface DerivedKeysResponse {
-    success: boolean;
-    data: DerivedKeys;
-    containerId: string;
-}
