+++ launchpad-starter-next-app/.env.example
@@ -1,3 +1,2 @@
 NEXT_PUBLIC_CROSSMINT_CLIENT_API_KEY=
-NEXT_PUBLIC_TEE_SERVER_URL=http://localhost:4000
 CROSSMINT_SERVER_API_KEY=
