+++ launchpad-starter-next-app/src/app/api/deploy/stop/route.ts
@@ -1,0 +1,22 @@
+import { NextResponse } from "next/server";
+import { ContainerManager } from "@/services/container";
+
+const containerManager = new ContainerManager();
+
+export async function POST() {
+    try {
+        await containerManager.stopContainer();
+        return NextResponse.json({
+            success: true,
+            message: "Agent stopped successfully",
+        });
+    } catch (error) {
+        return NextResponse.json(
+            {
+                success: false,
+                error: error instanceof Error ? error.message : "Unknown error",
+            },
+            { status: 500 }
+        );
+    }
+}
