+++ launchpad-starter-next-app/src/app/api/health/route.ts
@@ -1,0 +1,5 @@
+import { NextResponse } from "next/server";
+
+export function GET() {
+    return NextResponse.json({ status: "ok" });
+}
