+++ launchpad-starter-next-app/src/lib/utils.ts
@@ -1,3 +1,4 @@
+import { APIKeyEnvironmentPrefix, getEnvironmentForKey } from "@crossmint/common-sdk-base";
 import { type ClassValue, clsx } from "clsx";
 import { twMerge } from "tailwind-merge";
 
@@ -9,4 +10,15 @@ import { twMerge } from "tailwind-merge";
  */
 export function cn(...inputs: ClassValue[]) {
     return twMerge(clsx(inputs));
+}
+
+export function getBaseUrlFromApiKey(apiKey: string) {
+    switch (getEnvironmentForKey(apiKey)) {
+        case APIKeyEnvironmentPrefix.STAGING:
+            return "https://staging.crossmint.com/api/2022-06-09";
+        case APIKeyEnvironmentPrefix.PRODUCTION:
+            return "https://www.crossmint.com/api/2022-06-09";
+        default:
+            return "http://localhost:3000/api/2022-06-09";
+    }
 }
